chore: 공지사항 관련 파일 삭제
